% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_labels.R
\name{classify_labels}
\alias{classify_labels}
\title{Classify clear-text label name(s) into their parent (major-label) music labels}
\usage{
classify_labels(labels, concatenated = FALSE)
}
\arguments{
\item{labels}{A vector of clear-text (character) label names}

\item{concatenated}{Indicator, whether to return only a character vector with final classification (default is FALSE)}
}
\value{
Either a data frame of the same length as vector and three additional columns (sony, warner, universal), coded as 1 if clear-text label is part of one of the three major music labels, or not (0) (default). If \code{concatenated=TRUE},
returns a vector of same length as input vector.
}
\description{
Classify clear-text label name(s) into their parent (major-label) music labels
}
\examples{
classify_labels('Interscope')
classify_labels(c('Republic Records', 'Epic/Legacy', 'WM Finland', 'Chillhop Records'))
classify_labels(c('Republic Records', 'Epic/Legacy', 'WM Finland', 'Chillhop Records'), concatenated=T)
}
